first implementation of the common and repo lib to make it compile even if it is not perfect for now
