+++ crates/common/src/car/cc.rs
