+++ crates/common/src/car/fuel.rs
