+++ crates/common/src/car/price.rs
