+++ crates/common/src/lib.rs
@@ -1,3 +1,4 @@
-fn main() {
-    println!("Hello, world!");
-}
+pub mod car;
+pub mod error;
+pub mod intervention;
+pub mod std;
