+++ crates/common/src/std/email.rs
