+++ crates/repository/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+serde = { version = "1.0", features = ["derive"] }
+surrealdb = "2.0"
+thiserror = "1.0"
