+++ crates/repository/src/error.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum Error {
+    #[error("parsing error: {0}")]
+    Parsin(String),
+}
