+++ crates/repository/src/lib.rs
@@ -1,3 +1,2 @@
-fn main() {
-    println!("Hello, world!");
-}
+pub mod error;
+pub mod record;
