+++ crates/saas/Cargo.toml
@@ -23,3 +23,6 @@ surrealdb = "2.0"
 thiserror = "1.0"
 tokio = "1.40"
 tower-http = {version = "0.6.2", features= ["fs"]}
+
+common = {path="../common"}
+repository = {path="../repository"}
