+++ crates/saas/src/main.rs
@@ -1,4 +1,3 @@
-mod common;
 mod entity;
 mod error;
 mod handler;
